🎨 Feat: "Deatil 구조 개선"
Resolves: #9
