+++ src/components/RecommendList/RecommendListFood.vue
@@ -44,7 +44,7 @@
           </div>
         </div>
         <div>
-          <img alt="map" src="../assets/map.png" style="width:160px">
+          <img alt="map" src="@/assets/map.png" style="width:160px">
         </div>
       </div>
       <div class="content_list">
@@ -230,5 +230,5 @@ export default {
 </script>
 
 <style scoped>
-@import "../assets/recommendList_style.css";
+@import "@/assets/recommendList_style.css";
 </style>
