+++ src/router/index.js
@@ -4,11 +4,10 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
 import ReviewBoard from '../views/ReviewBoard.vue'
-// import RecommendFood from '../components/RecommendFood.vue' 
-import RecommendListFood from '../components/RecommendListFood.vue' 
-import RecommendListHotel from '../components/RecommendListHotel.vue' 
-import RecommendListPlace from '../components/RecommendListPlace.vue' 
-import RecommendListEvent from '../components/RecommendListEvent.vue' 
+import RecommendListFood from '../components/RecommendList/RecommendListFood.vue' 
+import RecommendListHotel from '../components/RecommendList/RecommendListHotel.vue' 
+import RecommendListPlace from '../components/RecommendList/RecommendListPlace.vue' 
+import RecommendListEvent from '../components/RecommendList/RecommendListEvent.vue' 
 import RecommendList from '../views/RecommendList.vue'
 //import RecommendList from '../views/RecommendListPlace.vue'
